requirements.txt
mammoth==1.11.0
markdown==3.10.0
pillow==10.4.0
playwright==1.48.0
python-docx==1.2.0
